requirements.txt
pymongo
sentence-transformers
fastapi
uvicorn
python-dotenv
streamlit
requests
pandas
pinecone